% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{LocusXcanR}
\alias{LocusXcanR}
\title{TWAS results integrated with other 'omics studies
R Shiny app to pull in relevant data for TWAS results interpretation}
\usage{
LocusXcanR(
  twas_result,
  weight_tbl,
  study_name = "",
  pred_exp_corr,
  conditional_present = FALSE,
  multiple_tissues = FALSE,
  known_variants,
  known_gwas,
  db_genes,
  all_gwas,
  ld_gwas,
  ref_expr_name = "",
  head_details = "",
  method_details = "",
  primary_tissue,
  meta_present = FALSE
)
}
\arguments{
\item{twas_result}{File path to TWAS results, as a text string (required)}

\item{weight_tbl}{File path to TWAS weights database, as a text string (required)}

\item{study_name}{A text string of the name of the study (optional, missing name is default)}

\item{pred_exp_corr}{File path to predicted expression correlation, as a text string (required)}

\item{conditional_present}{TRUE if conditional analysis results are available, FALSE if not (FALSE is default)}

\item{multiple_tissues}{TRUE if TWAS results are available from more than one tissue, FALSE if TWAS results are only available from a single tissue or a multi-tissue analysis (FALSE is default)}

\item{known_variants}{File path to known GWAS variants, as a text string (required)}

\item{known_gwas}{File path to study GWAS data matching known variants}

\item{db_genes}{File path to a list of genes in the database}

\item{all_gwas}{File path to study GWAS data}

\item{ld_gwas}{File path to the LD among study variants or an LD reference panel}

\item{ref_expr_name}{Name of the reference expression data set used in the analysis (optional, missing name is default)}

\item{head_details}{Any additional header details to be included in the app (optional, no details is default). HTML formatting commands may be used.}

\item{method_details}{Detailed methods section (optional, no details is default). HTLM formatting commands may be used.}

\item{primary_tissue}{If multiple tissues are present, list the name of the primary tissue}

\item{meta_present}{Results from TWAS meta-analysis present for comparison (optional, FALSE is default)}
}
\description{
created by: Amanda L. Tapia, Date: 06/08/2020
}
