% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{LocusXcanR}
\alias{LocusXcanR}
\title{LocusXcanR: An R Shiny App to Visualize TWAS Results Integrated with Other 'Omics Studies}
\usage{
LocusXcanR(
  twas_result,
  pvalthresh,
  weight_tbl,
  study_name = "",
  pred_exp_corr,
  conditional_present = FALSE,
  multiple_tissues = FALSE,
  known_variants,
  known_gwas,
  db_genes,
  all_gwas,
  ld_gwas,
  ref_expr_name = "",
  head_details = "",
  method_details = "",
  primary_tissue,
  meta_present = FALSE,
  meta_thresh,
  ideogram_present = FALSE,
  genome_build,
  cytoband_ds,
  add_UI = "",
  add_Server = ""
)
}
\arguments{
\item{twas_result}{character, file path to TWAS results (required). See Details for required column names.}

\item{pvalthresh}{numeric, -log10 p-value threshold for TWAS results (required).}

\item{weight_tbl}{character, file path to TWAS weights database (required). See Details for required column names.}

\item{study_name}{character, the name of the study (optional, default is missing)}

\item{pred_exp_corr}{character, file path to predicted expression correlation (required). See Details for required column names.}

\item{conditional_present}{logical, TRUE if conditional analysis results are available for plotting, FALSE otherwise (default is FALSE)}

\item{multiple_tissues}{logical, TRUE if TWAS results are available from more than one tissue, FALSE if TWAS results are only available from a single tissue or a multi-tissue analysis (default is FALSE)}

\item{known_variants}{character, file path to known GWAS variants (required). See Details for required column names.}

\item{known_gwas}{character, file path to study GWAS data matching known variants (required)}

\item{db_genes}{character, file path to a list of genes in the database (required)}

\item{all_gwas}{character, file path to study GWAS data (required)}

\item{ld_gwas}{character, file path to the LD among study variants or an LD reference panel (required)}

\item{ref_expr_name}{character, name of the reference expression data set used in the analysis (optional, default is missing)}

\item{head_details}{character, any additional header details to be included in the app (optional, default is no details). HTML formatting commands may be used.}

\item{method_details}{character, detailed methods section (optional, default is no details). HTLM formatting commands may be used.}

\item{primary_tissue}{character, if multiple tissues are available for analysis, list the name of the primary tissue (required if multiple_tissues = TRUE)}

\item{meta_present}{logical, TRUE if results from TWAS meta-analysis are present for comparison, FALSE otherwise (optional, default is FALSE)}

\item{meta_thresh}{numeric, p-value threshold for meta-analysis results (required if meta_present=TRUE)}

\item{ideogram_present}{logical, TRUE if an ideogramTrack should be plotted, FALSE otherwise (optional, default is FALSE)}

\item{genome_build}{character, the genome build for the ideogramTrack data (required if ideogram_present = TRUE)}

\item{cytoband_ds}{character, file path to the cytogenic bands data set (required if ideogram_present = TRUE)}

\item{add_UI}{character, additional elements to add to UI (optional, default is missing)}

\item{add_server}{character, additional elements to add to server (optional, default is missing)}
}
\value{
An R Shiny application
}
\description{
\code{LocusXcanR} allows users to import TWAS and GWAS results (and additional important details) 
in order to visualize them together in a single web page
application via R Shiny and facilitates TWAS results interpretation in context.
}
\details{
Please reference the vignette here () for further details on required and optional variables for each input dataset

Created by: Amanda L. Tapia, Date: June 8, 2020, Updated: December 16, 2020
}
\examples{

# This is an example with only the required parameters


# This is an example with all parameters


}
