% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{LocusXcanR}
\alias{LocusXcanR}
\title{LocusXcanR: An R Shiny App to Visualize TWAS Results Integrated with Other 'Omics Studies}
\usage{
LocusXcanR(
  twas_result,
  weight_tbl,
  study_name = "",
  pred_exp_corr,
  conditional_present = FALSE,
  multiple_tissues = FALSE,
  known_variants,
  known_gwas,
  db_genes,
  all_gwas,
  ld_gwas,
  ref_expr_name = "",
  head_details = "",
  method_details = "",
  primary_tissue,
  meta_present = FALSE,
  meta_thresh
)
}
\arguments{
\item{twas_result}{character, file path to TWAS results (required). See Details for required column names.}

\item{weight_tbl}{character, file path to TWAS weights database (required). See Details for required column names.}

\item{study_name}{character, the name of the study (optional, default is missing)}

\item{pred_exp_corr}{character, file path to predicted expression correlation (required). See Details for required column names.}

\item{conditional_present}{logical, TRUE if conditional analysis results are available for plotting, FALSE otherwise (default is FALSE)}

\item{multiple_tissues}{logical, TRUE if TWAS results are available from more than one tissue, FALSE if TWAS results are only available from a single tissue or a multi-tissue analysis (default is FALSE)}

\item{known_variants}{character, file path to known GWAS variants (required). See Details for required column names.}

\item{known_gwas}{character, file path to study GWAS data matching known variants}

\item{db_genes}{character, file path to a list of genes in the database}

\item{all_gwas}{character, file path to study GWAS data}

\item{ld_gwas}{character, file path to the LD among study variants or an LD reference panel}

\item{ref_expr_name}{character, name of the reference expression data set used in the analysis (optional, default is missing)}

\item{head_details}{character, any additional header details to be included in the app (optional, default is no details). HTML formatting commands may be used.}

\item{method_details}{character, detailed methods section (optional, default is no details). HTLM formatting commands may be used.}

\item{primary_tissue}{character, if multiple tissues are available for analysis, list the name of the primary tissue}

\item{meta_present}{logical, TRUE if results from TWAS meta-analysis are present for comparison, FALSE otherwise (optional, default is FALSE)}

\item{meta_thresh}{numeric, p-value threshold for meta-analysis results (required if meta_present=TRUE)}
}
\value{
An R Shiny application
}
\description{
\code{LocusXcanR} allows users to import TWAS and GWAS results (and additional important details) 
in order to visualize them together in a single web page
application via R Shiny and facilitates TWAS results interpretation in context.
}
\details{
This is the beginning of the details section

\code{twas_result} requires the following column names: 

\code{weight_tbl} requires the following column names: 

\code{known_variants} requires the following column names: 

\code{pred_exp_corr} requires the following column names: 

Created by: Amanda L. Tapia, Date: 06/08/2020
}
\examples{

This is an example of only the required stuff

This is an example with more stuff

}
